Add files via upload
This is a website centred around Dermatological products and services. It has a CV model which can differentiate between 5 skin diseases dataset fed to it when affected skin is presented in front of systems Camera. It also has an E-commerce webpage for beauty and dermatological products . Feature for appointment of doctors and dermatologists is added and data of customers and appointment schedule is stored real time at MongoDB database. Blogpost and recommendation pages have been added to enhance consumer experience.



Youtube video link of demonstration - https://youtu.be/CddN2dJ2QWg